Refactor DateTimeComparator to TemporalComparator to allow for low/upper udf customisation.
